backend/twelvelabs/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
httpx>=0.27.0
python-multipart>=0.0.9
pydantic>=2.6.0
python-dotenv>=1.0.0
twelvelabs>=0.3.0
deep-translator>=1.11.4
